+++ crates/sim/src/agents/portfolio_management/g3m/dca_g3m_liquidity_provider.rs
