+++ crates/sim/src/agents/portfolio_management/g3m/g3m_liquidity_provider.rs
