+++ crates/sim/src/agents/portfolio_management/g3m/mod.rs
