+++ crates/sim/src/agents/portfolio_management/lognormal/ln_arbitrageur.rs
@@ -8,7 +8,7 @@ use ethers::{
 };
 use tracing::info;
 
-use self::agents::portfolio_management_agents::base::arbitrageur::Arbitrageur;
+use self::agents::portfolio_management::base::arbitrageur::Arbitrageur;
 use super::{
     agent::*, agents::base::token_admin::TokenAdmin, Environment, Result, RevmMiddleware, *,
 };
